# Conflicts:
#	docs/usage.md

+++ src/v8unpack/v8unpack.py
@@ -11,27 +11,32 @@ from .json_container_decoder import json_decode, json_encode
 from .decoder import decode, encode
 from .file_organizer import FileOrganizer
 from . import helper
-from .index import create_index
+from .index import update_index
 from . import __version__
 
 
-def extract(in_filename: str, out_dir_name: str, *, temp_dir=None, index=None):
+def extract(in_filename: str, out_dir_name: str, *, temp_dir=None, index=None, version=None):
     begin0 = datetime.now()
     print(f"v8unpack {__version__}")
     print(f"{helper.str_time(begin0)} Начали        ", end='')
 
     helper.clear_dir(os.path.normpath(out_dir_name))
+    clear_temp_dir = False
     if temp_dir is None:
+        clear_temp_dir = True
         temp_dir = tempfile.mkdtemp()
-        helper.clear_dir(os.path.normpath(temp_dir))
+    helper.clear_dir(os.path.normpath(temp_dir))
 
     stage1_dir = os.path.join(temp_dir, 'decode_stage_1')
     stage2_dir = os.path.join(temp_dir, 'decode_stage_2')
     stage3_dir = os.path.join(temp_dir, 'decode_stage_3')
 
     if index:
-        with open(index, 'r', encoding='utf-8') as f:
-            index = json.load(f)
+        try:
+            with open(index, 'r', encoding='utf-8') as f:
+                index = json.load(f)
+        except FileNotFoundError:
+            index = None
 
     pool = helper.get_pool()
 
@@ -45,7 +50,7 @@ def extract(in_filename: str, out_dir_name: str, *, temp_dir=None, index=None):
 
     begin3 = datetime.now()
     print(f" - {begin3 - begin2}\n{helper.str_time(begin0)} Раcшифровываем", end='')
-    decode(stage2_dir, stage3_dir, pool=pool)
+    decode(stage2_dir, stage3_dir, pool=pool, version=version)
 
     begin4 = datetime.now()
     print(f" - {begin4 - begin3}\n{helper.str_time(begin0)} Организуем    ", end='')
@@ -55,16 +60,19 @@ def extract(in_filename: str, out_dir_name: str, *, temp_dir=None, index=None):
     print(f" - {end - begin4}\n{helper.str_time(end)} Готово         - {end - begin0}")
 
     helper.close_pool(pool)
-    shutil.rmtree(temp_dir, ignore_errors=True)
+    if clear_temp_dir:
+        shutil.rmtree(temp_dir, ignore_errors=True)
 
 
 def build(in_dir_name: str, out_file_name: str, *, temp_dir=None, index=None, version='803'):
     begin0 = datetime.now()
     print(f"v8unpack {__version__}")
     print(f"{helper.str_time(begin0)} Начали        ", end='')
+    clear_temp_dir = False
     if temp_dir is None:
+        clear_temp_dir = True
         temp_dir = tempfile.mkdtemp()
-        helper.clear_dir(os.path.normpath(temp_dir))
+    helper.clear_dir(os.path.normpath(temp_dir))
 
     encode_dir_stage1 = os.path.join(temp_dir, 'encode_stage_1')
     encode_dir_stage2 = os.path.join(temp_dir, 'encode_stage_2')
@@ -73,8 +81,11 @@ def build(in_dir_name: str, out_file_name: str, *, temp_dir=None, index=None, ve
     pool = helper.get_pool()
 
     if index:
-        with open(index, 'r', encoding='utf-8') as f:
-            index = json.load(f)
+        try:
+            with open(index, 'r', encoding='utf-8') as f:
+                index = json.load(f)
+        except FileNotFoundError:
+            index = None
 
     begin1 = datetime.now()
     print(f" - {begin1 - begin0}\n{helper.str_time(begin1)} Собираем      ", end='')
@@ -96,7 +107,8 @@ def build(in_dir_name: str, out_file_name: str, *, temp_dir=None, index=None, ve
     print(f" - {end - begin4}\n{helper.str_time(end)} Готово         - {end - begin0}")
 
     helper.close_pool(pool)
-    shutil.rmtree(temp_dir, ignore_errors=True)
+    if clear_temp_dir:
+        shutil.rmtree(temp_dir, ignore_errors=True)
 
 
 def main():
@@ -113,18 +125,18 @@ def main():
                        help='собрать файл 1С, где '
                             'src - путь до папки с исходниками, '
                             'file - путь до бинарного файла')
-    group.add_argument('-I', nargs=3, metavar=('file', 'src', 'dest'),
+    group.add_argument('-I', nargs=1, metavar='src',
                        help='сформировать index, где '
-                            'file - путь до файла с индексом, '
-                            'src - путь до папки с исходниками, '
-                            'dest - путь до общей папки')
+                            'src - путь до папки с исходниками'
+                       )
     parser.add_argument('--temp', help='путь до временной папки')
+    parser.add_argument('--core', help='название общей папки добавляемой в индекс по умолчанию')
     parser.add_argument('--index', help='путь до json файла с словарем копирования,'
                                         'структура файла: {путь исходника: путь общей папки}')
     parser.add_argument('--version', default='803',
                         help="версия сборки 801/802/803, по умолчанию 803, "
-                             " для сборки расширений указывается версия режима совместимости "
-                             "например для 8.3.6 это 80306, для разборки не требуется")
+                             " для сборки и разборки расширений указывается версия режима совместимости "
+                             "например для 8.3.6 это 80306, подробности в документации на github")
 
     if len(sys.argv) == 1:
         parser.print_help()
@@ -139,7 +151,7 @@ def main():
         build(args.B[0], args.B[1], index=args.index, temp_dir=args.temp, version=args.version)
 
     if args.I is not None:
-        create_index(args.I[0], args.I[1], args.I[2])
+        update_index(args.I[0], args.index, args.core)
 
 
 if __name__ == '__main__':
